+++ src/main/java/com/webeye/backend/cosmetic/infrastructure/persistence/init/IngredientInit.java
@@ -4,7 +4,6 @@ import com.webeye.backend.cosmetic.domain.Ingredient;
 import com.webeye.backend.cosmetic.domain.type.IngredientType;
 import com.webeye.backend.cosmetic.infrastructure.persistence.IngredientRepository;
 import com.webeye.backend.global.util.DummyDataInit;
-import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.ApplicationArguments;
@@ -19,7 +18,7 @@ public class IngredientInit implements ApplicationRunner {
 
     private final IngredientRepository ingredientRepository;
 
-    @Builder
+    @Override
     public void run(ApplicationArguments args) {
         for (IngredientType ingredientType : IngredientType.values()) {
             if (ingredientRepository.findByIngredientType(ingredientType).isEmpty()) {
